requirements.txt
pygame>=2.5.0
mutagen>=1.47.0
Pillow>=10.0.0
yt-dlp>=2024.1.1
spotdl>=4.2.0
